requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8

werkzeug~=0.16.0
pdf2image~=1.14.0
pytesseract~=0.3.7
pillow~=8.0.0
tqdm~=4.54.1
numpy~=1.19.2
